% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Codes to create grouped bar chart}
\usage{
barchart(
  data,
  y = "Prevalence",
  x = "Comorbidities",
  mygroup = "Severity",
  mycolor = c("#ABDDA4", "#66C2A5", "#3288BD"),
  mytitle = "Severity:"
)
}
\arguments{
\item{data}{example data for the manuscript is \code{comorbidities}}
}
\value{
a ggplot object
}
\description{
Codes to create grouped bar chart
}
\examples{

barchart <- barchart(data = comorbidities)
ggsave_custom("barchart.jpeg", imgpath = "~/charts/inst/img/",
inplot = barchart, device = jpeg)

}
