% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calculation.R
\name{calculate_log_rel_risk_bin}
\alias{calculate_log_rel_risk_bin}
\title{CI for log relative risk for binary outcomes}
\usage{
calculate_log_rel_risk_bin(prop1, prop2, N1, N2)
}
\arguments{
\item{prop1}{(\code{numeric})\cr Proportion of cases in active treatment}

\item{prop2}{(\code{numeric})\cr Proportion of cases in comparator treatment}

\item{N1}{(\code{numeric})\cr Total number of subjects in active treatment}

\item{N2}{(\code{numeric})\cr Total number of subjects in comparator treatment}
}
\description{
Derive log relative risk and associated confidence intervals
for binary outcomes
}
\examples{
calculate_log_rel_risk_bin(.45, 0.25, 500, 500)
}
