% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartsCircular.R
\name{ChartsCircular}
\alias{ChartsCircular}
\title{Circular Charts.}
\usage{
ChartsCircular(message, width = NULL, height = NULL)
}
\arguments{
\item{message}{A dataframe.}

\item{width, height}{nada.}
}
\value{
A circular charts based on ECharts 3.0.
}
\description{
Create a cricular chart based on ECharts 3.0.
}

