% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comorbidities}
\alias{comorbidities}
\title{Comorbidities data used for Figure 4}
\format{
A data frame with 15 rows and 3 variables
}
\usage{
comorbidities
}
\description{
Comorbidities data used for Figure 4
}
\keyword{datasets}
