% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormobidities.R
\name{cormobidities}
\alias{cormobidities}
\title{Codes to create cormobidities}
\usage{
cormobidities(data)
}
\arguments{
\item{data}{example data for the manuscript is \code{comorbidities}}
}
\value{
a ggplot object
}
\description{
Codes to create cormobidities
}
\examples{

cormobidities <- cormobidities(data = comorbidities)
ggsave_custom("cormobidities.jpeg", imgpath = "~/charts/inst/img/",
inplot = cormobidities, device = jpeg)

}
