% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demography}
\alias{demography}
\title{Example demography data}
\format{
\subsection{\code{who}}{

A data frame with 48 rows and 5 columns:
\describe{
\item{X}{Record ID number}
\item{Age}{Age in years}
\item{Type}{Diabetis Type A/B}
\item{Gender}{Gender Male/Female}
\item{Prevalence}{Prevalence}

}
}
}
\usage{
demography
}
\description{
Simulated demography data
}
\keyword{datasets}
