% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effects_table}
\alias{effects_table}
\title{Example effects table data used for Figure 6 and Figure 7}
\format{
A data frame with 24 rows and 51 variables
}
\usage{
effects_table
}
\description{
Example effects table data used for Figure 6 and Figure 7
}
\keyword{datasets}
