% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities.R
\name{figure4}
\alias{figure4}
\title{Codes to create comorbidities}
\usage{
figure4(data)
}
\arguments{
\item{data}{example data for the manuscript is \code{comorbidities}}
}
\value{
a ggplot object
}
\description{
Codes to create comorbidities
}
\examples{

fig4 <- figure4(data = comorbidities)
ggsave_custom("figure4.jpeg",
  imgpath = tempdir(),
  inplot = fig4, device = jpeg
)

}
