% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave_custom}
\alias{ggsave_custom}
\title{Wrapper to ggsave: Save a ggplot (or other grid object) with sensible
defaults}
\usage{
ggsave_custom(
  save_name,
  inplot,
  wdth = 7,
  hght = 4.1,
  unts = "in",
  imgpath = "inst/img/",
  bgcol = "white",
  create.dir = TRUE,
  ...
)
}
\arguments{
\item{save_name}{File name to create on disk.}

\item{inplot}{Plot to save, defaults to last plot displayed.}

\item{wdth, hght, unts}{Plot size in units ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{imgpath}{Path of the directory to save plot to: path}

\item{bgcol}{Background color. If NULL, uses the plot.background fill value
from the plot theme.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by device.}
}
\description{
Adds customized defaults to ggsave for the BRAP Journal requirements
}
\examples{
fig4 <- figure4(data = comorbidities)
ggsave_custom("figure4.jpeg", imgpath = "/cloud/project/inst/img/", inplot = fig4)
}
