% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_barchart.R
\name{grouped_barchart}
\alias{grouped_barchart}
\title{Grouped Bar Chart}
\usage{
grouped_barchart(data, xvar, yvar, groupvar, chartcolors)
}
\arguments{
\item{data}{\code{dataframe} input data}

\item{xvar}{\code{value} x-axis variable}

\item{yvar}{\code{value} y-axis variable}

\item{groupvar}{\code{Factor} group variable}

\item{chartcolors}{\code{vector} a vector of colors, the same number of levels as
the group variable}
}
\value{
a ggplot object
}
\description{
Grouped Bar Chart
}
\examples{

grouped_barchart(
  data = comorbidities, xvar = "Comorbidities",
  yvar = "Prevalence", groupvar = "Severity",
  chartcolors = colfun()$fig4_colors
)

}
