% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_chart.R
\name{pyramid_chart}
\alias{pyramid_chart}
\title{Pyramid Chart}
\usage{
pyramid_chart(data, xvar, yvar, levelvar, groupvar, alpha_set, chartcolors)
}
\arguments{
\item{data}{\code{dataframe} demography data}

\item{xvar}{\code{value} x-axis}

\item{yvar}{\code{value} y-axis}

\item{levelvar}{\code{Factor} two factor levels, one for each pyramid}

\item{groupvar}{\code{Factor} two factor levels, one for each side of a pyramid}

\item{alpha_set}{\code{Value} specify transparency of symbols}

\item{chartcolors}{\code{vector} two colors, one for each side of a pyramid}
}
\value{
ggplot object
}
\description{
Pyramid Chart
}
\examples{
demography |>
  dplyr::mutate(
    Type = as.factor(paste0("Type ", Type)),
    figprev = ifelse(
      Gender == "Females", -1 * Prevalence / 100000, Prevalence / 100000
    ),
    Sex = Gender
  ) |>
  pyramid_chart(
    levelvar = "Type", xvar = "figprev", yvar = "Age",
    groupvar = "Sex", alpha_set = 0.7, chartcolors = colfun()$fig2_colors
  )
}
