% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{pyramid_plot}
\alias{pyramid_plot}
\title{Demography Figure}
\usage{
pyramid_plot(
  fig2_data = demo_dset,
  Typev = c("A", "B"),
  gendervar = "Gender",
  alpha_set = 0.7
)
}
\arguments{
\item{fig2_data}{\code{dataframe} demography data}

\item{Typev}{\code{Factor} Type of Diabetis A or B}

\item{gendervar}{\code{Factor} Gender variable}
}
\value{
ggplot object
}
\description{
Demography Figure
}
\examples{
library(charts)
demography |>
  mutate(
    figprev = ifelse(
      Gender == "Females", -1 * Prevalence / 100000, Prevalence / 100000
    ),
    Sex = Gender
  ) |>
  pyramid_plot(gendervar = "Sex")
}
