% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_mermaid}
\alias{save_mermaid}
\title{Save DiagrameR::mermaid object}
\usage{
save_mermaid(diagfig, path)
}
\arguments{
\item{diagfig}{mermaid object}

\item{path}{\code{character} path to save file}
}
\value{
a saved image
}
\description{
Save DiagrameR::mermaid object
}
\examples{
if (interactive()) {
  # from github:
  remotes::install_github("bokeh/rbokeh")
  save_mermaid(value_tree(), paste0(tempdir(), "/value_tree.png"))
}
}
