% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barchart.R
\name{stacked_barchart}
\alias{stacked_barchart}
\title{Stacked Bar Chart}
\usage{
stacked_barchart(data, chartcolors, xlabel = "Visit")
}
\arguments{
\item{data}{\code{dataframe} input data}

\item{chartcolors}{\code{vector} a vector of colors, the same number of levels as
the brcat variable}
}
\value{
a ggplot object
}
\description{
Stacked Bar Chart
}
\examples{
stacked_barchart(
  data = comp_outcome,
  chartcolors = colfun()$fig12_colors,
  xlabel = "Study Week"
 )

}
