% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_tree.R
\name{value_tree}
\alias{value_tree}
\title{Create Value Tree}
\usage{
value_tree(
  diagram =
    "graph LR;\\n  A(<B>Benefit-Risk Balance</B>)-->B(<B>Benefits</B>)\\n  B-->C(<B>Primary Efficacy</B>)\\n  B-->D(<B>Secondary Efficacy</B>)\\n  B-->E(<B>Quality of life</B>)\\n  C-->F(<B>\% Success</B>)\\n  D-->G(<B>Mean change</B>)\\n  E-->H(<B>Mean change</B>)\\n  A-->I(<B>Risks</B>)\\n  I-->J(<B>Recurring AE</B>)\\n  I-->K(<B>Rare SAE</B>)\\n  I-->L(<B>Liver Toxicity</B>)\\n  J-->M(<B>Event rate</B>)\\n  K-->N(<B>\% Event</B>)\\n  L-->O(<B>\% Event</B>)\\n  style A fill:#7ABD7E\\n  style B fill:#7ABD7E\\n  style I fill:#7ABD7E\\n\\n  style C fill:#FFE733\\n  style D fill:#FFE733\\n  style E fill:#FFE733\\n  style J fill:#FFE733\\n  style K fill:#FFE733\\n  style L fill:#C6C6C6\\n\\n  style F fill: #FFAA1C\\n  style G fill: #FFAA1C\\n  style H fill: #FFAA1C\\n  style M fill: #FFAA1C\\n  style N fill: #FFAA1C\\n  style O fill: #C6C6C6\\n",
  ...
)
}
\value{
value tree image
}
\description{
Generate example value tree
}
\examples{
value_tree()
}
