% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comorbidities}
\alias{comorbidities}
\title{Example comorbidities data used for Figure 4}
\format{
A data frame with 15 rows and 3 variables
\describe{
\item{Comorbidities}{Comorbidities}
\item{Severity}{Severity of comorbidities Mild/Moderate/Severe}
\item{Prevalence}{Prevalence of comorbidities of each severity level}

}
}
\usage{
comorbidities
}
\description{
Example comorbidities data used for Figure 4
}
\keyword{datasets}
