% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_chart.R
\name{pyramid_chart}
\alias{pyramid_chart}
\title{Demography Figure}
\usage{
pyramid_chart(fig2_data, Typev, gendervar, alpha_set)
}
\arguments{
\item{fig2_data}{\code{dataframe} demography data}

\item{Typev}{\code{Factor} Type of Diabetis A or B}

\item{gendervar}{\code{Factor} Gender variable}

\item{alpha_set}{\code{Value} specify transparency of symbols}
}
\value{
ggplot object
}
\description{
Demography Figure
}
\examples{
library(charts)
demography |>
  dplyr::mutate(
    figprev = ifelse(
      Gender == "Females", -1 * Prevalence / 100000, Prevalence / 100000
    ),
    Sex = Gender
  ) |>
  pyramid_chart(Typev = c("A", "B"), gendervar = "Sex", alpha_set = 0.7)
}
