% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_chart.R
\name{pyramid_chart}
\alias{pyramid_chart}
\title{Pyramid Chart}
\usage{
pyramid_chart(data, xvar, yvar, levelsvar, gendervar, alpha_set)
}
\arguments{
\item{data}{\code{dataframe} demography data}

\item{xvar}{'Value' x-axis}

\item{yvar}{'value' y-axis}

\item{levelsvar}{\code{vector} factor levels}

\item{gendervar}{\code{Factor} Gender variable}

\item{alpha_set}{\code{Value} specify transparency of symbols}
}
\value{
ggplot object
}
\description{
Pyramid Chart
}
\examples{
library(charts)
demography |>
  dplyr::mutate(
    figprev = ifelse(
      Gender == "Females", -1 * Prevalence / 100000, Prevalence / 100000
    ),
    Sex = Gender
  ) |>
  pyramid_chart(levelsvar = c("A", "B"), x = "figprev", y = "Age",
  gendervar = "Sex", alpha_set = 0.7)
}
