% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_function.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot}
\usage{
scatter_plot(
  df_diff,
  outcome,
  legend_position = c(0, 1.05),
  fig_colors = c("#00AFBB", "red", "blue")
)
}
\arguments{
\item{df_diff}{A dataframe containing two vectors, each of which display the
difference between incremental probabilities in active and control effects
for a specified outcome.}

\item{outcome}{A vector of two strings that describes the two outcomes
associated with the difference in active and control effects, where the first
outcome corresponds to \code{diff1} and the second to \code{diff2}.}

\item{legend_position}{Allows user to specify legend position. Must be a
vector of length 2, with the first numeric value corresponding to the
position of the legend relative to the x-axis, and the second numeric value
corresponding to the position of the legend relative to the y-axis (defaults
are given).}

\item{fig_colors}{Allows user to change colors of the figure (defaults are
provided). Must be a vector of length 3, with the first color corresponding
to the scatter plot points, the second corresponding to the overall mean, and
third to the written probability text color.}
}
\value{
ggplot object
}
\description{
Scatter plot
}
\examples{
outcome <- c("Benefit", "Risk")
scatter_plot(scatterplot, outcome)
}
