% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barchart.R
\name{stacked_barchart}
\alias{stacked_barchart}
\title{Stacked Bar Chart}
\usage{
stacked_barchart(data, chartcolors, xlabel = "Visit")
}
\arguments{
\item{data}{\code{dataframe} input data}

\item{chartcolors}{\code{vector} a vector of colors, the same number of levels as
the brcat variable}

\item{xlabel}{\code{character} x label name, default is "Visit"}
}
\value{
a ggplot object
}
\description{
Stacked Bar Chart
}
\examples{
library(testthat)
plot1 <- stacked_barchart(
  data = comp_outcome,
  chartcolors = colfun()$fig12_colors,
  xlabel = "Study Week"
 )
ggsave(plot1, filename = paste0(test_path(),"/snapshots/stacked_barchart1.png"), width = 7, height = 5)

# unequal number of subjects across treatments
comp_outcome2 <- comp_outcome[
  (comp_outcome$trtn == 1 & comp_outcome$usubjid \%in\% c(1:40))
  | (comp_outcome$trtn == 2 & comp_outcome$usubjid \%in\% c(101:160))
  | (comp_outcome$trtn == 3 & comp_outcome$usubjid \%in\% c(201:250))
,]
plot2 <- stacked_barchart(
  data = comp_outcome2,
  chartcolors = colfun()$fig12_colors,
  xlabel = "Study Week"
 )
ggsave(plot2, filename = paste0(test_path(),"/snapshots/stacked_barchart2.png"), width = 7, height = 5)

# unequal number of observations across visits
comp_outcome3 <- comp_outcome[!(comp_outcome$trtn == 1 &
  comp_outcome$usubjid \%in\% c(1:40) & comp_outcome$visit == 5), ]
plot3 <- stacked_barchart(
  data = comp_outcome3,
  chartcolors = colfun()$fig12_colors,
  xlabel = "Study Week"
 )
ggsave(plot3, filename = paste0(test_path(),"/snapshots/stacked_barchart3.png"), width = 7, height = 5)
}
